Moved non-gui code for bluetooth widget
Moved non-gui related code for the bluetooth widget into it own module class
